+++ util/CHANGELOG.md
@@ -5,6 +5,10 @@ Refer to https://keepachangelog.com/en/1.0.0/ for how to maintain changelog<br /
 
 ## [Unreleased]
 
+-
+
+## [0.3.0] - 03/05/2023
+
 -   add customisation of namespace colour for logger
 -   improve error stack formatting for logger
 -   flatten export structure